Move code that depends on k8s libs outside of client/pps
