+++ src/server/pps/util.go
@@ -2,6 +2,11 @@ package pps
 
 import (
 	"fmt"
+	"math"
+	"strings"
+
+	"k8s.io/kubernetes/pkg/api"
+	kube "k8s.io/kubernetes/pkg/client/unversioned"
 
 	"github.com/pachyderm/pachyderm/src/client/pfs"
 	ppsclient "github.com/pachyderm/pachyderm/src/client/pps"
@@ -15,4 +20,52 @@ func JobRepo(job *ppsclient.Job) *pfs.Repo {
 // PipelineRepo creates a pfs repo for a given pipeline.
 func PipelineRepo(pipeline *ppsclient.Pipeline) *pfs.Repo {
 	return &pfs.Repo{Name: pipeline.Name}
+}
+
+// PipelineRcName generates the name of the k8s replication controller that
+// manages a pipeline's workers
+func PipelineRcName(name string, version uint64) string {
+	// k8s won't allow RC names that contain upper-case letters
+	// or underscores
+	// TODO: deal with name collision
+	name = strings.Replace(name, "_", "-", -1)
+	return fmt.Sprintf("pipeline-%s-v%d", strings.ToLower(name), version)
+}
+
+// GetExpectedNumWorkers computes the expected number of workers that
+// pachyderm will start given the ParallelismSpec 'spec'.
+//
+// This is only exported for testing
+func GetExpectedNumWorkers(kubeClient *kube.Client, spec *ppsclient.ParallelismSpec) (int, error) {
+	coefficient := 0.0 // Used if [spec.Strategy == PROPORTIONAL] or [spec.Constant == 0]
+	if spec == nil {
+		// Unset ParallelismSpec is handled here. Currently we start one worker per
+		// node
+		coefficient = 1.0
+	} else if spec.Strategy == ppsclient.ParallelismSpec_CONSTANT {
+		if spec.Constant > 0 {
+			return int(spec.Constant), nil
+		}
+		// Zero-initialized ParallelismSpec is handled here. Currently we start one
+		// worker per node
+		coefficient = 1
+	} else if spec.Strategy == ppsclient.ParallelismSpec_COEFFICIENT {
+		coefficient = spec.Coefficient
+	} else {
+		return 0, fmt.Errorf("Unable to interpret ParallelismSpec strategy %s", spec.Strategy)
+	}
+	if coefficient == 0.0 {
+		return 0, fmt.Errorf("Ended up with coefficient == 0 (no workers) after interpreting ParallelismSpec %s", spec.Strategy)
+	}
+
+	// Start ('coefficient' * 'nodes') workers. Determine number of workers
+	nodeList, err := kubeClient.Nodes().List(api.ListOptions{})
+	if err != nil {
+		return 0, fmt.Errorf("unable to retrieve node list from k8s to determine parallelism: %v", err)
+	}
+	if len(nodeList.Items) == 0 {
+		return 0, fmt.Errorf("pachyderm.pps.jobserver: no k8s nodes found")
+	}
+	result := math.Floor(coefficient * float64(len(nodeList.Items)))
+	return int(math.Max(result, 1)), nil
 }
